+++ pacifica-core/src/main/java/com/trs/pacifica/log/file/AbstractFile.java
@@ -49,7 +49,7 @@ public abstract class AbstractFile {
 
     private final String filename;
 
-    private final int fileSize;
+    private final long fileSize;
 
     private long lastLogIndex;
